api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
openai==1.40.1
httpx==0.27.2
pydantic==2.7.1
python-dotenv==1.0.1
chromadb==0.5.3
neo4j==5.28.0
python-multipart==0.0.9
tiktoken==0.7.0
pyyaml>=6.0.0
langdetect==1.0.9
deep-translator==1.11.4
pytest==8.3.3
gTTS==2.5.1
elevenlabs==1.3.0
indic-transliteration==2.3.61
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator>=2.1.0
redis==5.0.1
hiredis==2.3.2